cleanup(sinsp): rename cri.cpp to cri.hpp
